Fix middleware bug
The bug occurs because Flask tries to register the middleware with the
name of the wrapper function "wrapper". It works once, but more than
once will cause a collision. By specifying that the wrapper wraps a
specific function, we can avoid this problem

Solution in part from https://stackoverflow.com/a/42254713